+++ sped_cfe/models/inherited_sped_documento.py
@@ -362,46 +362,6 @@ class SpedDocumento(models.Model):
 
         return cliente
 
-    def processador_vfpe(self):
-        """
-        Busca classe do processador do cadastro da empresa, onde podemos
-        ter três tipos de processamento dependendo
-        de onde o equipamento esta instalado:
-
-        - Instalado no mesmo servidor que o Odoo;
-        - Instalado na mesma rede local do servidor do Odoo;
-        - Instalado em um local remoto onde o browser vai ser responsável
-        por se comunicar com o equipamento
-
-        :return:
-        """
-        self.ensure_one()
-
-        cliente = None
-
-        if self.configuracoes_pdv.tipo_sat == 'local':
-            from mfecfe import BibliotecaSAT
-            from mfecfe import ClienteVfpeLocal
-
-            chave = self.configuracoes_pdv.chave_acesso_validador
-            cliente = ClienteVfpeLocal(
-                BibliotecaSAT(
-                    self.configuracoes_pdv.path_integrador),
-                chave_acesso_validador=chave,
-                numerador_sessao=self._gera_numero_sessao(),
-            )
-        elif self.configuracoes_pdv.tipo_sat == 'rede_interna':
-            from mfecfe.clientesathub import ClienteVfpeHub
-            cliente = ClienteVfpeHub(
-                self.configuracoes_pdv.ip,
-                self.configuracoes_pdv.porta,
-                numero_caixa=int(self.configuracoes_pdv.numero_caixa)
-            )
-        elif self.configuracoes_pdv.tipo_sat == 'remoto':
-            raise NotImplementedError
-
-        return cliente
-
     def grava_cfe(self, cfe):
         self.ensure_one()
         nome_arquivo = 'envio-cfe.xml'
@@ -904,64 +864,3 @@ class SpedDocumento(models.Model):
                 pagamentos_cartoes.append(pagamento)
 
         return pagamentos_cartoes
-
-    def envia_pagamento(self):
-        self.ensure_one()
-        pagamentos_cartoes = self._verificar_formas_pagamento()
-        if not pagamentos_cartoes:
-            self.pagamento_autorizado_cfe = True
-        else:
-            pagamentos_autorizados = True
-            config = self.configuracoes_pdv
-            cliente = self.processador_vfpe()
-
-            for duplicata in pagamentos_cartoes:
-                if not duplicata.id_pagamento:
-                    if self.configuracoes_pdv.tipo_sat == 'local':
-                        resposta = cliente.enviar_pagamento(
-                            config.chave_requisicao, duplicata.estabecimento,
-                            duplicata.serial_pos, config.cnpjsh,
-                            self.bc_icms_proprio,
-                            duplicata.valor,
-                            config.multiplos_pag,
-                            config.anti_fraude,
-                            'BRL',
-                            int(config.numero_caixa),
-                        )
-                    elif self.configuracoes_pdv.tipo_sat == 'rede_interna':
-                        resposta = cliente.enviar_pagamento(
-                            config.chave_requisicao,
-                            duplicata.estabecimento,
-                            duplicata.serial_pos,
-                            config.cnpjsh,
-                            self.bc_icms_proprio,
-                            duplicata.valor,
-                            config.multiplos_pag,
-                            config.anti_fraude,
-                            'BRL',
-                            int(config.numero_caixa),
-                            config.chave_acesso_validador,
-                            config.path_integrador,
-                        )
-                    resposta_pagamento = resposta.split('|')
-                    if len(resposta_pagamento[0]) >= 7:
-                        duplicata.id_pagamento = resposta_pagamento[0]
-                        duplicata.id_fila = resposta_pagamento[1]
-                    else:
-                        pagamentos_autorizados = False
-                # FIXME status sempre vai ser negativo na homologacao
-                # resposta_status_pagamento =
-                        # cliente.verificar_status_validador(
-                #     config.cnpjsh, duplicata.id_fila_status
-                # )
-                #
-                # resposta_status_pagamento =
-                        # cliente.verificar_status_validador(
-                #     config.cnpjsh, '214452'
-                # )
-                # if resposta_status_pagamento.ValorPagamento
-                        # == '0' and resposta_status_pagamento.IdFila == '0':
-                #     pagamentos_autorizados = False
-                #     break
-
-            self.pagamento_autorizado_cfe = pagamentos_autorizados
